# Optimize heartbeat intervals for faster leader convergence